# ARROW-9029: [C++] Implement BitBlockCounter for much faster block popcounts of bitmaps