--- a/cpp/src/arrow/util/bit_util.h
+++ b/cpp/src/arrow/util/bit_util.h
@@ -1170,11 +1170,18 @@ class BitsetStack {
   std::vector<int> offsets_;
 };
 
-/// \brief A class that scans through a true/false bitmap to yield runs of up
+/// \brief A class that scans through a true/false bitmap to yield blocks of up
 /// to 256 bits at a time along with their popcount. This is used to accelerate
 /// processing of mostly-not-null array data.
 class ARROW_EXPORT BitmapScanner {
  public:
+  struct Block {
+    int16_t length;
+    int16_t popcount;
+  };
+
+  static constexpr int16_t kTargetBlockLength = 256;
+
   BitmapScanner(const uint8_t* bitmap, int64_t start_offset, int64_t length)
       : bitmap_(bitmap + start_offset / 8),
         bits_remaining_(length),
@@ -1182,7 +1189,7 @@ class ARROW_EXPORT BitmapScanner {
 
   /// \brief Return the next run of available bits, up to 256. The returned
   /// pair contains the size of run and the number of true values
-  std::pair<int64_t, int64_t> NextRun();
+  Block NextBlock();
 
  private:
   const uint8_t* bitmap_;