--- a/cpp/src/arrow/util/bit_util.h
+++ b/cpp/src/arrow/util/bit_util.h
@@ -1170,10 +1170,11 @@ class BitsetStack {
   std::vector<int> offsets_;
 };
 
-/// \brief A class that scans through a true/false bitmap to yield runs of
-/// all-true or not-all-true runs. This is used to accelerate processing of
-/// mostly-not-null array data. 4 words are examined at a time.
-struct ARROW_EXPORT BitmapScanner {
+/// \brief A class that scans through a true/false bitmap to yield runs of up
+/// to 256 bits at a time along with their popcount. This is used to accelerate
+/// processing of mostly-not-null array data.
+class ARROW_EXPORT BitmapScanner {
+ public:
   BitmapScanner(const uint8_t* bitmap, int64_t start_offset, int64_t length)
       : bitmap_(bitmap + start_offset / 8),
         bits_remaining_(length),
@@ -1183,6 +1184,7 @@ struct ARROW_EXPORT BitmapScanner {
   /// pair contains the size of run and the number of true values
   std::pair<int64_t, int64_t> NextRun();
 
+ private:
   const uint8_t* bitmap_;
   int64_t bits_remaining_;
   int64_t offset_;