--- a/cpp/src/arrow/util/bit_util_benchmark.cc
+++ b/cpp/src/arrow/util/bit_util_benchmark.cc
@@ -22,10 +22,12 @@
 #include <bitset>
 #include <vector>
 
+#include "arrow/array/array_base.h"
 #include "arrow/buffer.h"
 #include "arrow/builder.h"
 #include "arrow/memory_pool.h"
 #include "arrow/testing/gtest_util.h"
+#include "arrow/testing/random.h"
 #include "arrow/testing/util.h"
 #include "arrow/util/bit_util.h"
 
@@ -387,6 +389,114 @@ static void BitmapEquals(benchmark::State& state) {
   state.SetBytesProcessed(state.iterations() * buffer_size);
 }
 
+template <int64_t Offset = 0>
+static void BitmapScannerSumNotNull(benchmark::State& state) {
+  using internal::BitmapScanner;
+
+  random::RandomArrayGenerator rng(/*seed=*/0);
+
+  const int64_t bitmap_length = 1 << 20;
+
+  // State parameter is the average number of total values for each false
+  // value. So 100 means that 1 out of 100 on average are false.
+  double true_probability = 1. - 1. / state.range(0);
+  auto arr = rng.Int8(bitmap_length, 0, 100, true_probability);
+
+  const uint8_t* bitmap = arr->null_bitmap_data();
+
+  // Compute the expected result
+  int64_t expected = 0;
+  const auto& int8_arr = static_cast<const Int8Array&>(*arr);
+  for (int64_t i = Offset; i < bitmap_length; ++i) {
+    if (int8_arr.IsValid(i)) {
+      expected += int8_arr.Value(i);
+    }
+  }
+  for (auto _ : state) {
+    BitmapScanner scanner(bitmap, Offset, bitmap_length - Offset);
+    int64_t result = 0;
+    int64_t position = Offset;
+    while (true) {
+      auto bit_run = scanner.NextRun();
+      if (bit_run.first == 0) {
+        break;
+      }
+      if (bit_run.first == bit_run.second) {
+        // All not-null
+        for (int64_t i = 0; i < bit_run.first; ++i) {
+          result += int8_arr.Value(position + i);
+        }
+      } else if (bit_run.second > 0) {
+        // Some but not all not-null
+        for (int64_t i = 0; i < bit_run.first; ++i) {
+          if (BitUtil::GetBit(bitmap, position + i)) {
+            result += int8_arr.Value(position + i);
+          }
+        }
+      }
+      position += bit_run.first;
+    }
+    // Sanity check
+    if (result != expected) {
+      std::abort();
+    }
+  }
+  state.SetItemsProcessed(state.iterations() * bitmap_length);
+}
+
+template <int64_t Offset = 0>
+static void BitmapReaderSumNotNull(benchmark::State& state) {
+  random::RandomArrayGenerator rng(/*seed=*/0);
+
+  const int64_t bitmap_length = 1 << 20;
+
+  // State parameter is the average number of total values for each false
+  // value. So 100 means that 1 out of 100 on average are false.
+  double true_probability = 1. - 1. / state.range(0);
+  auto arr = rng.Int8(bitmap_length, 0, 100, true_probability);
+
+  const uint8_t* bitmap = arr->null_bitmap_data();
+  // Compute the expected result
+  int64_t expected = 0;
+  const auto& int8_arr = static_cast<const Int8Array&>(*arr);
+  for (int64_t i = Offset; i < bitmap_length; ++i) {
+    if (int8_arr.IsValid(i)) {
+      expected += int8_arr.Value(i);
+    }
+  }
+  for (auto _ : state) {
+    internal::BitmapReader bit_reader(bitmap, Offset, bitmap_length - Offset);
+    int64_t result = 0;
+    for (int64_t i = Offset; i < bitmap_length; ++i) {
+      if (bit_reader.IsSet()) {
+        result += int8_arr.Value(i);
+      }
+      bit_reader.Next();
+    }
+    // Sanity check
+    if (result != expected) {
+      std::abort();
+    }
+  }
+  state.SetItemsProcessed(state.iterations() * bitmap_length);
+}
+
+static void BitmapScannerSumNotNull(benchmark::State& state) {
+  BitmapScannerSumNotNull<0>(state);
+}
+
+static void BitmapScannerSumNotNullWithOffset(benchmark::State& state) {
+  BitmapScannerSumNotNull<4>(state);
+}
+
+static void BitmapReaderSumNotNull(benchmark::State& state) {
+  BitmapReaderSumNotNull<0>(state);
+}
+
+static void BitmapReaderSumNotNullWithOffset(benchmark::State& state) {
+  BitmapReaderSumNotNull<4>(state);
+}
+
 static void BitmapEqualsWithoutOffset(benchmark::State& state) { BitmapEquals<0>(state); }
 
 static void BitmapEqualsWithOffset(benchmark::State& state) { BitmapEquals<4>(state); }
@@ -423,6 +533,12 @@ BENCHMARK(CopyBitmapWithOffsetBoth)->Arg(kBufferSize);
 
 BENCHMARK(BitmapEqualsWithoutOffset)->Arg(kBufferSize);
 BENCHMARK(BitmapEqualsWithOffset)->Arg(kBufferSize);
+
+// Range value: average number of total values per null
+BENCHMARK(BitmapScannerSumNotNull)->Range(8, 1 << 16);
+BENCHMARK(BitmapScannerSumNotNullWithOffset)->Range(8, 1 << 16);
+BENCHMARK(BitmapReaderSumNotNull)->Range(8, 1 << 16);
+BENCHMARK(BitmapReaderSumNotNullWithOffset)->Range(8, 1 << 16);
 
 #define AND_BENCHMARK_RANGES                      \
   {                                               \