@wesm 2020-06-03
@cyb70289 since you've been doing bitmap stuff lately I would welcome your code review of this