@wesm 2020-06-04
Thanks for the comments, I think this is much better now