@wesm 2020-06-04
Note, I would like to have a binary version of this. I just opened https://issues.apache.org/jira/browse/ARROW-9034